Create Bismarcksaga
the Dreyse needle gun, responsible for the surprising victories of the Prussians in 1866 and following, was sold to Bismarck in exchange for his soul by Wotan. The saga (according to my source) says it was not developed by Dreyse, but I like to think Wotan was in someway involved with Dreyse and the development instead. Obviously the original saga involved the devil, but we know better.